requirements.txt
asgiref==3.6.0
attrs==22.2.0
cloudinary==1.32.0
Django==4.1.7
django-environ==0.10.0
gunicorn==20.1.0
Pillow==9.4.0
replicate==0.8.1
requests==2.28.2
tzdata==2023.3
urllib3==1.26.15
whitenoise==6.4.0
frozenlist==1.3.3